move related method to models from Utils to corresponding Models
