+++ app/src/main/java/com/github/elianaferreira/movieslist/stories/detail/movie/Videos.kt
@@ -1,4 +1,5 @@
-package com.github.elianaferreira.movieslist.models
+package com.github.elianaferreira.movieslist.stories.detail.movie
+
 import androidx.annotation.Keep
 import com.google.gson.annotations.SerializedName
 
@@ -9,28 +10,37 @@ data class Videos(
     val id: Int,
     @SerializedName("results")
     val results: List<Video>
-)
+) {
+    fun getTrailerKey(): String {
+        //get only trailer
+        val trailers = results.filter { it.type.lowercase() == "trailer" }
+        if (trailers.isNotEmpty()) {
+            return trailers.first().key
+        } else {
+            //if there is no trailer, search for teaser
+            val teasers = results.filter { it.type.lowercase() == "teaser" }
+            return if (teasers.isNotEmpty()) {
+                teasers.first().key
+            } else {
+                //search for clips
+                val clips = results.filter { it.type.lowercase() == "clip" }
+                if (clips.isNotEmpty()) {
+                    clips.first().key
+                } else {
+                    //extreme case: return the first element's key
+                    results.first().key
+                }
+            }
+        }
+    }
+}
 
 @Keep
 data class Video(
     @SerializedName("id")
     val id: String,
-    @SerializedName("iso_3166_1")
-    val iso31661: String,
-    @SerializedName("iso_639_1")
-    val iso6391: String,
     @SerializedName("key")
     val key: String,
-    @SerializedName("name")
-    val name: String,
-    @SerializedName("official")
-    val official: Boolean,
-    @SerializedName("published_at")
-    val publishedAt: String,
-    @SerializedName("site")
-    val site: String,
-    @SerializedName("size")
-    val size: Int,
     @SerializedName("type")
     val type: String
 )
