+++ app/src/main/java/com/github/elianaferreira/movieslist/stories/home/Category.kt
@@ -1,9 +1,14 @@
 package com.github.elianaferreira.movieslist.stories.home
 
 import java.io.Serializable
+import java.util.*
 
 data class Category (
     val categoryValue: String,
     val categoryName: String,
     val drawable: Int
-) : Serializable
+) : Serializable {
+    fun categoryIsMovie() : Boolean {
+        return categoryValue.lowercase(Locale.getDefault()).contains("movie")
+    }
+}
