+++ app/src/main/java/com/github/elianaferreira/movieslist/utils/RequestManager.kt
@@ -10,8 +10,8 @@ import com.android.volley.toolbox.Volley
 import com.github.elianaferreira.movieslist.BuildConfig
 import com.github.elianaferreira.movieslist.R
 import com.github.elianaferreira.movieslist.stories.list.MoviesList
-import com.github.elianaferreira.movieslist.models.*
 import com.github.elianaferreira.movieslist.stories.detail.movie.MovieDetail
+import com.github.elianaferreira.movieslist.stories.detail.movie.Videos
 import com.github.elianaferreira.movieslist.stories.detail.tvshow.TVShowDetail
 import com.google.gson.Gson
 import java.lang.StringBuilder
